test(contract): assert user input, assistant output, and tool calls on chat spans
Move the message-content helpers (_text_parts, _tool_call_parts) into
GenAITestBase and assert on parsed chat-span messages there so all three
gen AI framework tests get the coverage: a user input, an assistant
output, and a tool call in the aggregated chat-span input messages.

The OpenAI Agents test keeps its invoke_agent content assertions (first
user input + last agent output), which only that instrumentor emits.
